refactor code structor
- move `onHost.hpp` -> `onHost/interface.hpp`
- move `onAcc.hpp` to `onAcc/interface.hpp`
- move `*/internal.hpp` to */internal/interface.hpp`

This refactoring should make it easier for the user to know which folders should be inspected to find all interfaces of alpaka.
