+++ include/alpaka/alpaka.hpp
@@ -19,18 +19,17 @@
 #include "alpaka/core/config.hpp"
 #include "alpaka/executor.hpp"
 #include "alpaka/interface.hpp"
-#include "alpaka/internal.hpp"
+#include "alpaka/internal/interface.hpp"
 #include "alpaka/math.hpp"
 #include "alpaka/math/constants.hpp"
 #include "alpaka/math/internal/Complex.hpp"
 #include "alpaka/mem/Iter.hpp"
-#include "alpaka/onAcc.hpp"
 #include "alpaka/onAcc/Acc.hpp"
 #include "alpaka/onAcc/GlobalMem.hpp"
 #include "alpaka/onAcc/SimdAlgo.hpp"
 #include "alpaka/onAcc/atomic.hpp"
+#include "alpaka/onAcc/interface.hpp"
 #include "alpaka/onAcc/tag.hpp"
-#include "alpaka/onHost.hpp"
 #include "alpaka/onHost/Device.hpp"
 #include "alpaka/onHost/DeviceSelector.hpp"
 #include "alpaka/onHost/Queue.hpp"
@@ -39,6 +38,7 @@
 #include "alpaka/onHost/algo/reduce.hpp"
 #include "alpaka/onHost/algo/transform.hpp"
 #include "alpaka/onHost/algo/transformReduce.hpp"
+#include "alpaka/onHost/interface.hpp"
 #include "alpaka/onHost/mem/stdContainer.hpp"
 #include "alpaka/tag.hpp"
 #include "utility.hpp"
