+++ include/alpaka/api/host/atomic.hpp
@@ -8,7 +8,7 @@
 #include "alpaka/core/config.hpp"
 #include "alpaka/onAcc/atomicHierarchy.hpp"
 #include "alpaka/onAcc/atomicOp.hpp"
-#include "alpaka/onAcc/internal.hpp"
+#include "alpaka/onAcc/internal/interface.hpp"
 
 #include <array>
 #include <atomic>
