+++ include/alpaka/api/oneApi/Platform.hpp
@@ -9,9 +9,9 @@
 #if ALPAKA_LANG_ONEAPI
 #    include "alpaka/api/oneApi/Api.hpp"
 #    include "alpaka/api/syclGeneric/Platform.hpp"
-#    include "alpaka/internal.hpp"
+#    include "alpaka/internal/interface.hpp"
 #    include "alpaka/onHost/example/executors.hpp"
-#    include "alpaka/onHost/internal.hpp"
+#    include "alpaka/onHost/internal/interface.hpp"
 #    include "alpaka/tag.hpp"
 
 namespace alpaka
