+++ include/alpaka/api/syclGeneric/Event.hpp
@@ -11,9 +11,9 @@
 #    include "alpaka/api/util.hpp"
 #    include "alpaka/core/CallbackThread.hpp"
 #    include "alpaka/interface.hpp"
-#    include "alpaka/internal.hpp"
+#    include "alpaka/internal/interface.hpp"
 #    include "alpaka/onHost/concepts.hpp"
-#    include "alpaka/onHost/internal.hpp"
+#    include "alpaka/onHost/internal/interface.hpp"
 
 #    include <sycl/sycl.hpp>
