+++ include/alpaka/api/syclGeneric/Queue.hpp
@@ -13,11 +13,11 @@
 #    include "alpaka/api/util.hpp"
 #    include "alpaka/core/CallbackThread.hpp"
 #    include "alpaka/interface.hpp"
-#    include "alpaka/internal.hpp"
+#    include "alpaka/internal/interface.hpp"
 #    include "alpaka/onAcc/Acc.hpp"
-#    include "alpaka/onHost.hpp"
 #    include "alpaka/onHost/concepts.hpp"
-#    include "alpaka/onHost/internal.hpp"
+#    include "alpaka/onHost/interface.hpp"
+#    include "alpaka/onHost/internal/interface.hpp"
 #    include "alpaka/onHost/mem/ManagedView.hpp"
 #    include "alpaka/onHost/trait.hpp"
