+++ include/alpaka/api/syclGeneric/atomic.hpp
@@ -11,7 +11,7 @@
 #    include "alpaka/api/syclGeneric/tag.hpp"
 #    include "alpaka/onAcc/atomicHierarchy.hpp"
 #    include "alpaka/onAcc/atomicOp.hpp"
-#    include "alpaka/onAcc/internal.hpp"
+#    include "alpaka/onAcc/internal/interface.hpp"
 
 #    include <sycl/sycl.hpp>
