+++ include/alpaka/api/unifiedCudaHip/atomic.hpp
@@ -10,7 +10,7 @@
 #include "alpaka/core/config.hpp"
 #include "alpaka/onAcc/atomicHierarchy.hpp"
 #include "alpaka/onAcc/atomicOp.hpp"
-#include "alpaka/onAcc/internal.hpp"
+#include "alpaka/onAcc/internal/interface.hpp"
 
 #include <limits>
 #include <type_traits>
